requirements.txt
aiofiles==23.2.1
alembic>=1.11.3
axabc>=0.0.64
aiogram==3.4.1
aiogram-dialog==2.1.0
aiosqlite==0.20.0
axsqlalchemy==0.0.69
dishka==0.8.0
fastapi==0.110.1
httpx==0.27.0
python-dotenv==1.0.1
SQLAlchemy==2.0.3
uvicorn==0.29.0
yarl==1.9.4
pydantic_settings
grateful_logging
aiohttp-socks
minio==7.2.7
python-multipart
python-jose
passlib
pydantic[email]
